Deleted Previous files and added the whole projects that is fully operational but with only one thing still needed to be taken care of
Here I would talk about the dispencery that had arised. We have input rows to take the user internship and experience details but they have not been added to the latex file generating python code tex_gen.py which is the only thing needed to be taken care of.
Further, we cannot modify the line stucture of base.txt since we are using the lines numbers in few python files.